Abstruct the metric capture code
