+++ app/models/manageiq/providers/kubernetes/container_manager/metrics_capture/capture_context_mixin.rb
@@ -1,7 +1,5 @@
 class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
-  class CaptureContext
-    include ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture::HawkularClientMixin
-
+  module CaptureContextMixin
     def initialize(target, start_time, end_time, interval)
       @target = target
       @start_time = start_time || 15.minutes.ago.beginning_of_minute.utc
@@ -12,11 +10,11 @@ class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
       @ts_values = Hash.new { |h, k| h[k] = {} }
       @metrics = []
 
-      if @target.respond_to?(:hardware)
-        hardware = @target.hardware
-      else
-        hardware = @target.try(:container_node).try(:hardware)
-      end
+      hardware = if @target.respond_to?(:hardware)
+                   @target.hardware
+                 else
+                   @target.try(:container_node).try(:hardware)
+                 end
 
       @node_cores = hardware.try(:cpu_total_cores)
       @node_memory = hardware.try(:memory_mb)
@@ -54,75 +52,12 @@ class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
       raise TargetValidationError, "Validation error: memory not defined" unless @node_memory.to_i > 0
     end
 
-    def collect_node_metrics
-      cpu_resid = "machine/#{@target.name}/cpu/usage"
-      process_cpu_counters_rate(fetch_counters_rate(cpu_resid))
-
-      mem_resid = "machine/#{@target.name}/memory/usage"
-      process_mem_gauges_data(fetch_gauges_data(mem_resid))
-
-      net_resid = "machine/#{@target.name}/network"
-      net_counters = [fetch_counters_rate("#{net_resid}/tx"),
-                      fetch_counters_rate("#{net_resid}/rx")]
-
-      process_net_counters_rate(compute_summation(net_counters))
-    end
-
-    def collect_container_metrics
-      group_id = @target.container_group.ems_ref
-
-      cpu_resid = "#{@target.name}/#{group_id}/cpu/usage"
-      process_cpu_counters_rate(fetch_counters_rate(cpu_resid))
-
-      mem_resid = "#{@target.name}/#{group_id}/memory/usage"
-      process_mem_gauges_data(fetch_gauges_data(mem_resid))
-    end
-
-    def collect_group_metrics
-      group_id = @target.ems_ref
-
-      cpu_counters = @target.containers.collect do |c|
-        fetch_counters_rate("#{c.name}/#{group_id}/cpu/usage")
-      end
-      process_cpu_counters_rate(compute_summation(cpu_counters))
-
-      mem_gauges = @target.containers.collect do |c|
-        fetch_gauges_data("#{c.name}/#{group_id}/memory/usage")
-      end
-      process_mem_gauges_data(compute_summation(mem_gauges))
-
-      net_resid = "pod/#{group_id}/network"
-      net_counters = [fetch_counters_rate("#{net_resid}/tx"),
-                      fetch_counters_rate("#{net_resid}/rx")]
-      process_net_counters_rate(compute_summation(net_counters))
-    end
-
     def fetch_counters_rate(resource)
       compute_derivative(fetch_counters_data(resource))
     end
 
-    def fetch_counters_data(resource)
-      sort_and_normalize(
-        hawkular_client.counters.get_data(
-          resource,
-          :starts         => (@start_time - @interval).to_i.in_milliseconds,
-          :bucketDuration => "#{@interval}s"))
-    rescue SystemCallError, SocketError, OpenSSL::SSL::SSLError => e
-      raise CollectionFailure, e.message
-    end
-
-    def fetch_gauges_data(resource)
-      sort_and_normalize(
-        hawkular_client.gauges.get_data(
-          resource,
-          :starts         => @start_time.to_i.in_milliseconds,
-          :bucketDuration => "#{@interval}s"))
-    rescue SystemCallError, SocketError, OpenSSL::SSL::SSLError => e
-      raise CollectionFailure, e.message
-    end
-
     def process_cpu_counters_rate(counters_rate)
-      @metrics |= ['cpu_usage_rate_average'] if counters_rate.length > 0
+      @metrics |= ['cpu_usage_rate_average'] unless counters_rate.empty?
       total_cpu_time = @node_cores * CPU_NANOSECONDS * @interval
       counters_rate.each do |x|
         timestamp = Time.at(x['start'] / 1.in_milliseconds).utc
@@ -132,7 +67,7 @@ class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
     end
 
     def process_mem_gauges_data(gauges_data)
-      @metrics |= ['mem_usage_absolute_average'] if gauges_data.length > 0
+      @metrics |= ['mem_usage_absolute_average'] unless gauges_data.empty?
       gauges_data.each do |x|
         timestamp = Time.at(x['start'] / 1.in_milliseconds).utc
         avg_usage = (x['avg'] / 1.megabytes) * 100.0 / @node_memory
@@ -141,7 +76,7 @@ class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
     end
 
     def process_net_counters_rate(counters_rate)
-      @metrics |= ['net_usage_rate_average'] if counters_rate.length > 0
+      @metrics |= ['net_usage_rate_average'] unless counters_rate.empty?
       counters_rate.each do |x|
         timestamp = Time.at(x['start'] / 1.in_milliseconds).utc
         avg_usage_kb = x['avg'] / (1.kilobyte.to_f * @interval)
@@ -165,13 +100,6 @@ class ManageIQ::Providers::Kubernetes::ContainerManager::MetricsCapture
           }
         end
       end
-    end
-
-    def sort_and_normalize(data)
-      # Sorting and removing last entry because always incomplete
-      # as it's still in progress.
-      norm_data = (data.sort_by { |x| x['start'] }).slice(0..-2)
-      norm_data.reject { |x| x.values.include?('NaN') || x['empty'] == true }
     end
 
     def compute_derivative(counters)
